Restructured files, made backend able to serve React client and added ESLint static style, error and type checking
