+++ client/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Ludum",
+  "name": "Ludum",
   "icons": [
     {
       "src": "favicon.ico",
@@ -11,5 +11,5 @@
   "start_url": ".",
   "display": "standalone",
   "theme_color": "#26c6da",
-  "background_color": "#ffffff"
+  "background_color": "#f7f7f7"
 }
