--- client/src/components/AccountButton/AccountButton.scss
@@ -1,55 +1,0 @@
-@import '@material/react-button/index.scss';
-@import '@material/react-icon-button/index.scss';
-@import '@material/react-menu-surface/index.scss';
-@import '@material/react-typography/index.scss';
-.account {
-    margin: auto 0;
-    img {
-        border-radius: 50%;
-        height: 48px;
-        width: 48px;
-    }
-}
-
-.mdc-menu-surface {
-    padding: 8px;
-    display: grid;
-    grid-row-gap: 16px;
-    grid-template-areas: 
-    "name name name name" 
-    "email email email email" 
-    "profile . . ."
-    "sign-out . dark-mode settings";
-}
-
-#name {
-    grid-area: name;
-}
-
-#email {
-    margin-top: -16px;
-    grid-area: email;
-}
-
-#profile {
-    grid-area: profile;
-    @include mdc-button-container-fill-color(var(--mdc-theme-secondary));
-}
-
-#dark-mode {
-    grid-area: dark-mode;
-    margin: auto;
-}
-
-#settings {
-    grid-area: settings;
-    margin: auto 0 0 auto;
-}
-
-#sign-out {
-    grid-area: sign-out;
-    margin: auto;
-    color: var(--mdc-theme-secondary);
-    @include mdc-button-outline-color(var(--mdc-theme-secondary));
-    @include mdc-states-base-color(#b00020);
-}
