--- client/src/components/MainLayout/MainLayout.scss
@@ -1,35 +1,0 @@
-@import '@material/react-drawer/index.scss';
-@import '@material/react-list/index.scss';
-@import '@material/react-top-app-bar/index.scss';
-@import '@material/react-material-icon/index.scss';
-.drawer-container {
-    display: flex;
-    flex-direction: row;
-    height: 100vh;
-    overflow: hidden;
-}
-
-.mdc-drawer {
-    @include mdc-drawer-surface-fill-color(var(--mdc-theme-surface));
-}
-
-.drawer-app-content {
-    flex: auto;
-    overflow: auto;
-}
-
-.mdc-top-app-bar__section--align-end {
-    padding-right: 267px; /* This padding is required because of the side-drawer. */
-}
-
-#navList .mdc-list-item {
-    font-size: 1rem;
-}
-
-/* Dark theming. */
-html[data-theme="dark"] { 
-    .mdc-drawer {
-        @include mdc-drawer-fill-color-accessible(#006064);
-        @include mdc-drawer-item-activated-icon-ink-color(#da5026);
-    }
-}
