--- client/src/components/MainLayout/desktop.ini
@@ -1,5 +1,0 @@
-[.ShellClassInfo]
-InfoTip=Den här mappen delas online.
-IconFile=C:\Program Files\Google\Drive\googledrivesync.exe
-IconIndex=16
-
