+++ client/src/components/common/App.scss
@@ -1,0 +1,93 @@
+@import '@material/react-drawer/index.scss';
+@import '@material/react-top-app-bar/index.scss';
+@import '@material/react-drawer/index.scss';
+@import '@material/react-list/index.scss';
+@import '@material/react-button/index.scss';
+@import '@material/react-icon-button/index.scss';
+@import '@material/react-menu-surface/index.scss';
+.drawer-container {
+    display: flex;
+    flex-direction: row;
+    height: 100vh;
+    overflow: hidden;
+}
+
+.drawer-app-content {
+    flex: auto;
+    overflow: auto;
+}
+
+.mdc-top-app-bar__section--align-end {
+    /* This padding is required because of the side-drawer. */
+    padding-right: 267px;
+}
+
+.mdc-drawer {
+    @include mdc-drawer-surface-fill-color(var(--mdc-theme-surface));
+}
+
+#navList .mdc-list-item {
+    font-size: 1rem;
+}
+
+/* Dark theming. */
+
+html[data-theme="dark"] {
+    .mdc-drawer {
+        @include mdc-drawer-fill-color-accessible(#006064);
+        /* Should always be --mdc-theme-surface color. */
+        @include mdc-drawer-item-activated-icon-ink-color(#da5026);
+    }
+}
+
+.account {
+    margin: auto 0;
+    img {
+        border-radius: 50%;
+        height: 48px;
+        width: 48px;
+    }
+}
+
+.mdc-menu-surface #grid {
+    padding: 8px;
+    display: grid;
+    grid-row-gap: 16px;
+    grid-template-areas: "name name name name" "email email email email" "profile . . ." "sign-out . dark-mode settings";
+}
+
+#name {
+    grid-area: name;
+}
+
+#email {
+    margin-top: -16px;
+    grid-area: email;
+}
+
+#profile {
+    grid-area: profile;
+    @include mdc-button-container-fill-color(var(--mdc-theme-secondary));
+}
+
+#dark-mode {
+    grid-area: dark-mode;
+    margin: auto;
+    width: 48px;
+    height: 48px;
+}
+
+#settings {
+    grid-area: settings;
+    margin: auto 0 0 auto;
+    width: 48px;
+    height: 48px;
+}
+
+#sign-out {
+    grid-area: sign-out;
+    margin: auto;
+    color: var(--mdc-theme-secondary);
+    @include mdc-button-outline-color(var(--mdc-theme-secondary));
+    @include mdc-states-base-color(#b00020);
+}
