+++ client/src/styles.scss
@@ -1,3 +1,6 @@
+@import '@material/react-material-icon/index.scss';
+@import '@material/react-typography/index.scss';
+
 * {
     margin: 0;
     padding: 0;
