--- package.json
@@ -1,15 +1,0 @@
-{
-  "name": "ludum",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "concurrently \"cd backend && npm start\" \"cd client && npm start\""
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "concurrently": "^4.0.1"
-  }
-}
